backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
duckduckgo-search
newspaper3k
lxml
lxml_html_clean
chromadb
sentence-transformers
